Allow leap tool to have limited uses (Issue #22)
This adds a new config.yml variable (tools.leap.cost) which allows for
for a limited number of leaps before one of the leap tool items is
removed from the users inventory. By default it is set to zero, which
means there is no limit to the number of uses (old behavior is default).

There is also a new permission node (toolbelt.tool.leap.free) which
allows a player to bypass the cost if there is one configured in the
config.yml file.