Stop warnings for depreciation of material.Door
org.bukkit.material.Door was deprecated with Bukkit 1.4.5-R1.0

This was because of the Minecraft 1.2 update where door data was spread
 between the top and bottom half of doors. I've already updated my
plugin to account for this and so material.Door still performs as
desired.

If/When Bukkit gets a new interface for handling doors this will be
removed and the code updated to use the new interface.